Implemented updating via PUT requests to /[collection]/[primary key]
The user can update an object by sending a PUT requset to the url of that object.

The standard behavior is to perform a merge update, overwriting or adding only
those fields provided in the body of the PUT request.

The user may change this behavior by supplying overwrite=true in the query string.
This causes the object to be overwritten by the object in the request body,
clobbering whatever is in the database.

The API returns the new object when the request completes, or an error if one occurs.

It might be desirable to return a 304 if no change occurs, but I see no simple
way to accomplish this at the moment.

close #1