Expanded Pagination
fixes #7

Instead of a page variable, the user now supplies skip and limit variables via
query string on both GET to /[collection] and POST to /[collection]/query.

These variables are then passed to the mongo cursor, causing it to skip the
first [skip] results and only return a maximum of [limit] results.

So to implement pagination, let limit = pageSize and skip = (page-1)*pageSize